requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
numpy==1.24.3
opencv-python==4.8.0.76
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.21
tensorflow==2.13.1